Build: Add support for GCC 5.0+
JIRA: MADLIB-1025

GCC 5.1 release libstdc++ introduced a new library ABI that includes new
implementations of std::string and std::list. We disable this ABI to
ensure MADlib compiles with non-gcc compilers as well.

Further, using -O3 optimization (default in CMAKE_BUILD_TYPE=Release)
led to runtime errors. We revert to O2 optimization level (default in
CMAKE_BUILD_TYPE=RelWithDebInfo) to avoid this issue.
The RCA for this problem is yet unknown.

+++ src/modules/convex/type/state.hpp
@@ -669,8 +669,11 @@ private:
             reinterpret_cast<dimension_pointer_type>(&mStorage[1]);
         task.stepsize.rebind(&mStorage[N + 2]);
         task.lambda.rebind(&mStorage[N + 3]);
-        uint32_t sizeOfModel = task.model.rebind(&mStorage[N + 4],&mStorage[N + 5],&mStorage[N + 6],
-                task.numberOfStages, task.numbersOfUnits);
+        size_t sizeOfModel = task.model.rebind(&mStorage[N + 4],
+                                                 &mStorage[N + 5],
+                                                 &mStorage[N + 6],
+                                                 task.numberOfStages,
+                                                 task.numbersOfUnits);
 
         algo.incrModel.rebind(&mStorage[N + 4],&mStorage[N + 5],&mStorage[N + 6 + sizeOfModel],
                 task.numberOfStages, task.numbersOfUnits);
